fix: resolve perps deposit max issues
